+++ src/common/constants.lua
@@ -3,7 +3,8 @@ local constants = {}
 constants.MAX_UNDERNAME_LENGTH = 61
 constants.UNDERNAME_DOES_NOT_EXIST_MESSAGE = "Name does not exist in the ANT!"
 constants.INVALID_ARWEAVE_ID_MESSAGE = "Invalid Arweave ID"
-constants.MIN_TTL_SECONDS = 900
+constants.MIN_TTL_SECONDS = 60
+constants.DEFAULT_TTL_SECONDS = 900
 constants.MAX_TTL_SECONDS = 3600
 constants.INVALID_TTL_MESSAGE = "Invalid TTL. TLL must be an integer between "
 	.. constants.MIN_TTL_SECONDS